requirements.txt
Django>=5.2
djangorestframework>=3.13
whitenoise>=6.0  # Make sure this line is present
gunicorn>=20.1